구현이 필요한 인터페이스들을 위한 TestConfiguration 정의
